day 1 - update
환경 정리 및 업데이트
- 백준 2309 (python) 풀이

+++ 코드업/basic-c/1044.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-    int a;
-    scanf("%d", &a);
+    long long int a;
+    scanf("%lld", &a);
     // a + 1
     // ++a 
-    printf("%d \n", ++a);
+    printf("%lld \n", ++a);
 }
